requirements.txt
requests==2.32.3
pandas==2.2.3
confluent-kafka==2.5.0
kafka-python==2.0.2
